:sparkles: expiredTime타입 변경, Id값 변경

+++ src/main/kotlin/com/example/todolist/global/security/jwt/properties/JwtExpTimeProperties.kt
@@ -6,6 +6,6 @@ import org.springframework.boot.context.properties.ConstructorBinding
 @ConstructorBinding
 @ConfigurationProperties("jwt.time")
 class JwtExpTimeProperties(
-    val accessExp: Long,
-    val refreshExp: Long
+    val accessExp: Int,
+    val refreshExp: Int
 )
